Selecting more than one item in the hooks dialog and calling "Remove..." only deleted every second item. The reason for this is that the index changes if an item is deleted. Solution: the list is now traversed from the end to the beginning
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@7426 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/SetHooks.cpp
@@ -118,16 +118,20 @@ void CSetHooks::RebuildHookList()
 
 void CSetHooks::OnBnClickedRemovebutton()
 {
-	POSITION pos = m_cHookList.GetFirstSelectedItemPosition();
-	while (pos)
+	// traversing from the end to the beginning so that the indices are not skipped
+	int index = m_cHookList.GetItemCount()-1;
+	while (index >= 0)
 	{
-		int index = m_cHookList.GetNextSelectedItem(pos);
-		hookkey key;
-		key.htype = CHooks::GetHookType((LPCTSTR)m_cHookList.GetItemText(index, 0));
-		key.path = CTSVNPath(m_cHookList.GetItemText(index, 1));
-		CHooks::Instance().Remove(key);
-		m_cHookList.DeleteItem(index);
-		SetModified();
+		if (m_cHookList.GetItemState(index, LVIS_SELECTED) & LVIS_SELECTED)
+		{
+			hookkey key;
+			key.htype = CHooks::GetHookType((LPCTSTR)m_cHookList.GetItemText(index, 0));
+			key.path = CTSVNPath(m_cHookList.GetItemText(index, 1));
+			CHooks::Instance().Remove(key);		
+			m_cHookList.DeleteItem(index);
+			SetModified();
+		}
+		index--;
 	}
 }
